# api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API for SHOP Website",
    "description": "This is a REST API application made by ZDiDane for SHOP Website. \n\nIf you want to use this API, please get csrfToken first, then add csrfToken to header of request. \n\nIf you want to use API for Admin, please add bearAuth to header of request.",
    "contact": {
      "name": "ZDiDane",
      "email": "didannguyen17@gmail.com",
      "url": "https://didan.id.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    },
    {
      "url": "https://localhost:3000/api"
    },
    {
      "url": "http://shop.didan.id.vn/api"
    },
    {
      "url": "https://shop.didan.id.vn/api"
    }
  ],
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get all products",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/find-product": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to find product",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Product name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/cart": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get cart",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}